requirements.txt
pytest-playwright
python-dotenv
twilio
playwright-stealth
setuptools